requirements.txt
pytz==2020.1
Django==3.1
django-configurations==2.2
gunicorn==20.0.4
django-heroku

# Rest apis
djangorestframework==3.11.1
Markdown==3.2.2
django-filter==2.3.0

# For the persistence stores
psycopg2-binary==2.8.5
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.0.0
django_unique_upload==0.2.1